+++ .eslintrc.cjs
@@ -1,61 +1,54 @@
 /** @type {import("eslint").Linter.Config} */
 const config = {
-  "parser": "@typescript-eslint/parser",
-  "parserOptions": {
-    "project": true
+  parser: '@typescript-eslint/parser',
+  parserOptions: {
+    project: true,
   },
-  "plugins": [
-    "@typescript-eslint",
-    "drizzle"
+  plugins: ['@typescript-eslint', 'drizzle'],
+  extends: [
+    'next/core-web-vitals',
+    'plugin:@typescript-eslint/recommended-type-checked',
+    'plugin:@typescript-eslint/stylistic-type-checked',
   ],
-  "extends": [
-    "next/core-web-vitals",
-    "plugin:@typescript-eslint/recommended-type-checked",
-    "plugin:@typescript-eslint/stylistic-type-checked"
-  ],
-  "rules": {
-    "@typescript-eslint/array-type": "off",
-    "@typescript-eslint/consistent-type-definitions": "off",
-    "@typescript-eslint/consistent-type-imports": [
-      "warn",
+  rules: {
+    '@typescript-eslint/array-type': 'off',
+    '@typescript-eslint/consistent-type-definitions': 'off',
+    '@typescript-eslint/prefer-nullish-coalescing': 'off',
+    '@typescript-eslint/no-inferrable-types': 'off',
+    '@typescript-eslint/consistent-type-imports': [
+      'warn',
       {
-        "prefer": "type-imports",
-        "fixStyle": "inline-type-imports"
-      }
+        prefer: 'type-imports',
+        fixStyle: 'inline-type-imports',
+      },
     ],
-    "@typescript-eslint/no-unused-vars": [
-      "warn",
+    '@typescript-eslint/no-unused-vars': [
+      'warn',
       {
-        "argsIgnorePattern": "^_"
-      }
+        argsIgnorePattern: '^_',
+      },
     ],
-    "@typescript-eslint/require-await": "off",
-    "@typescript-eslint/no-misused-promises": [
-      "error",
+    '@typescript-eslint/require-await': 'off',
+    '@typescript-eslint/no-misused-promises': [
+      'error',
       {
-        "checksVoidReturn": {
-          "attributes": false
-        }
-      }
+        checksVoidReturn: {
+          attributes: false,
+        },
+      },
     ],
-    "drizzle/enforce-delete-with-where": [
-      "error",
+    'drizzle/enforce-delete-with-where': [
+      'error',
       {
-        "drizzleObjectName": [
-          "db",
-          "ctx.db"
-        ]
-      }
+        drizzleObjectName: ['db', 'ctx.db'],
+      },
     ],
-    "drizzle/enforce-update-with-where": [
-      "error",
+    'drizzle/enforce-update-with-where': [
+      'error',
       {
-        "drizzleObjectName": [
-          "db",
-          "ctx.db"
-        ]
-      }
-    ]
-  }
-}
+        drizzleObjectName: ['db', 'ctx.db'],
+      },
+    ],
+  },
+};
 module.exports = config;
